+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Changed
+
+- Move finalizer from AWSCluster to Cluster
+
 ### Added
 
 - Add CAPA-created VPC and security group IDs for usage with the Cilium ENI mode feature (to add pod network security groups via Crossplane)
+- Support for EKS AWSManagedControlPlane
 
 ## [0.1.1] - 2024-03-26